requirements.txt.txt
Flask==1.1.2
google-generativeai